+++ api/lib/schema.js
@@ -340,6 +340,11 @@ const typeDefs = gql`
     properties: String # JSON encoded as it is general purpose and GraphQL doesn't allow for "Object"
   }
 
+  type Session {
+    sessionToken: String!
+    validUntil: String!
+  }
+
   type Subscription {
     blockAdded(networkId: String!): BlockV2
     notificationAdded(addressObjects: [NotificationInput]!): Notification
@@ -349,10 +354,11 @@ const typeDefs = gql`
   }
 
   type Mutation {
-    registerUser(idToken: String!): Boolean
+    registerUser(idToken: String!): Session
     notifications(
       addressObjects: [NotificationInput]!
       notificationType: String
+      pushToken: String
     ): Boolean
   }